+++ xnumon.c
@@ -16,8 +16,6 @@
 #include "kextctl.h"
 #include "build.h"
 #include "policy.h"
-#include "auclass.h"
-#include "aupolicy.h"
 
 #include <stdlib.h>
 #include <stdio.h>
@@ -233,42 +231,9 @@ main(int argc, char *argv[]) {
 		goto errout;
 	}
 
-	/* system-global configuration, only run this if holding pidfile */
-	int pol = AUDIT_ARGV;
-	if (cfg->envlevel > 0)
-		pol |= AUDIT_ARGE;
-	if (aupolicy_ensure(pol) == -1) {
-		fprintf(stderr, "Failed to configure audit policy\n");
-		goto errout;
-	}
-	if (auclass_addmask(AC_XNUMON, auclass_xnumon_events_procmon) == -1) {
-		fprintf(stderr, "Failed to configure AC_XNUMON class mask\n");
-		goto errout;
-	}
-	if (LOGEVT_WANT(cfg->events, LOGEVT_HACKMON) &&
-	    auclass_addmask(AC_XNUMON, auclass_xnumon_events_hackmon) == -1) {
-		fprintf(stderr, "Failed to configure AC_XNUMON class mask\n");
-		goto errout;
-	}
-	if (LOGEVT_WANT(cfg->events, LOGEVT_FILEMON) &&
-	    auclass_addmask(AC_XNUMON, auclass_xnumon_events_filemon) == -1) {
-		fprintf(stderr, "Failed to configure AC_XNUMON class mask\n");
-		goto errout;
-	}
-
 	rv = evtloop_run(cfg);
 	if (rv == -1) {
 		fprintf(stderr, "Event loop returned error\n");
-	}
-
-	/* this is system-global, only run this if holding pidfile */
-	if (auclass_removemask(AC_XNUMON,
-	                       auclass_xnumon_events_procmon) == -1 ||
-	    auclass_removemask(AC_XNUMON,
-	                       auclass_xnumon_events_hackmon) == -1 ||
-	    auclass_removemask(AC_XNUMON,
-	                       auclass_xnumon_events_filemon) == -1) {
-		fprintf(stderr, "Failed to configure AC_XNUMON class mask\n");
 	}
 
 errout: